re-write with new features and removing the Redis driver
The process has changed to include one line in the
Console\Kernel::schedule method “OClock::register($schedule);” and it
will set the hooks for before and after and record the output in a file
that will be read in the “after” hook and stores it in the session’s
document then removes the file.

- Remove the Redis driver, only use MongoDB
- Written more tests for the components (Schedule, Event, Session)
- OClock is now a facade and methods can be called on the facade
directly and it aliases itself automatically in the service provider
- Removed the Report command for it’s no longer needed
- Introduced Sessions, every run of an event is a Session and it stores
all its details in there.